Add first pass of omniauth support for Shibboleth auth
When using the development environment, users will have a choice of
login: Shibboleth or the normal login page. In the production
environment, Shibboleth authentication will be automatic.

Note: This commit also adds a middleware component for development
that adds a REMOTE_USER request header that would normally be created
as part of normal authentication via Shibboleth & Apache.